The Whole Zip of Data-Visualization-Website
This is the whole zip of our critical_skills' project---Data-Visualization-Website
There are three main folder in the zip which are "back"(the web backend part), "front"(the web frontend part) and "database".
Our website has already been uploaded to the serve and could be log in at any time.
To be honest, i'm the leader of the backend part. Allow me to introduce some key words of that:
By using "Request" , we could make it crawl the data from National Statistical Database automatically. Through libs like "pandas","to_mysql" ,we could analyse and tabulate the data we have and upload it to "mysql", which could be used to make charts.
It was a happy and substantial time in my college life. I would like to use this library to commemorate that joyful and fulfilling time.